pyproject.toml
[project]
name = "team-task-manager"
version = "0.1.0"
description = "Team Task Management System with Django and Telegram Bot"
authors = [
    {name = "ALEXSTONE", email = "solo-drive@bk.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Django>=4.2.7,<5.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "djangorestframework>=3.14.0,<4.0.0",
    "aiogram>=3.2.0,<4.0.0",
    "requests (>=2.32.5,<3.0.0)",
    "channels (>=4.3.1,<5.0.0)",
    "channels-redis (>=4.3.0,<5.0.0)",
    "daphne (>=4.2.1,<5.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"